+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Change log
 ================================================================================
 
-0.6.2 - tbd
+0.6.2 - 15.09.2019
 --------------------------------------------------------------------------------
 
 Added
@@ -9,6 +9,10 @@ Added
 
 #. `#322 <https://github.com/moremoban/moban/issues/322>`_: Implicit targets
    with template extensions default to copy
+#. `#257 <https://github.com/moremoban/moban/issues/257>`_: '-e' to load
+   extensions for template engines, i.e. jinja2
+#. `#333 <https://github.com/moremoban/moban/issues/333>`_: command line
+   template fails with version 0.6.1
 
 0.6.1 - 10.09.2019
 --------------------------------------------------------------------------------